split Statistics.tsx into Totals.tsx and Titlecard.tsx
